refactor: DTO request/response 분리 (#24)
