+++ src/main/java/com/umc/hwaroak/dto/request/FriendRequestDto.java
@@ -1,4 +1,4 @@
-package com.umc.hwaroak.dto;
+package com.umc.hwaroak.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
