+++ src/main/java/com/umc/hwaroak/dto/response/KakaoUserInfoDto.java
@@ -1,7 +1,7 @@
 // 카카오 access token으로 사용자 정보를 가져올 때 응답받기 위한 dto
 // 클라이언트에게 주는 응답아님
 
-package com.umc.hwaroak.dto;
+package com.umc.hwaroak.dto.response;
 
 import lombok.Getter;
 import lombok.Setter;
